Exp-58: add validation for input "block" data to be integer, not numberic

+++ src/validations/TreatmentValidator.js
@@ -129,7 +129,7 @@ class TreatmentValidator extends SchemaValidator {
         keys: ['experimentId', 'treatmentNumber'],
         entity: db.treatment,
       },
-      { paramName: 'block', type: 'numeric' },
+      { paramName: 'block', type: 'integer' },
       { paramName: 'inAllBlocks', type: 'boolean' },
     ]
   }
